+++ integration/vite-css-test.ts
@@ -215,7 +215,7 @@ test.describe("Vite CSS", () => {
           cwd = await createProject(
             {
               "react-router.config.ts": reactRouterConfig({
-                viteEnvironmentApi: templateName !== "vite-5-template",
+                v8_viteEnvironmentApi: templateName !== "vite-5-template",
               }),
               "vite.config.ts": await viteConfig.basic({
                 port,
@@ -257,7 +257,7 @@ test.describe("Vite CSS", () => {
           cwd = await createProject(
             {
               "react-router.config.ts": reactRouterConfig({
-                viteEnvironmentApi: templateName !== "vite-5-template",
+                v8_viteEnvironmentApi: templateName !== "vite-5-template",
                 basename: base,
               }),
               "vite.config.ts": await viteConfig.basic({
@@ -300,7 +300,7 @@ test.describe("Vite CSS", () => {
           cwd = await createProject(
             {
               "react-router.config.ts": reactRouterConfig({
-                viteEnvironmentApi: templateName !== "vite-5-template",
+                v8_viteEnvironmentApi: templateName !== "vite-5-template",
               }),
               "vite.config.ts": await viteConfig.basic({
                 port,
@@ -342,7 +342,7 @@ test.describe("Vite CSS", () => {
           cwd = await createProject(
             {
               "react-router.config.ts": reactRouterConfig({
-                viteEnvironmentApi: templateName !== "vite-5-template",
+                v8_viteEnvironmentApi: templateName !== "vite-5-template",
               }),
               "vite.config.ts": await viteConfig.basic({
                 port,
@@ -413,7 +413,7 @@ test.describe("Vite CSS", () => {
           cwd = await createProject(
             {
               "react-router.config.ts": reactRouterConfig({
-                viteEnvironmentApi: templateName !== "vite-5-template",
+                v8_viteEnvironmentApi: templateName !== "vite-5-template",
               }),
               "vite.config.ts": await viteConfig.basic({
                 port,