+++ packages/react-router/__tests__/router/data-strategy-test.ts
@@ -32,6 +32,27 @@ describe("router dataStrategy", () => {
     );
   }
 
+  function keyedResultsUsingShouldCallHandler(
+    matchesToLoad: DataStrategyMatch[],
+    results: DataStrategyResult[],
+  ) {
+    if (matchesToLoad.length !== results.length) {
+      throw new Error(
+        `Mismatched results length: expected ${matchesToLoad.length} but got ${results.length}`,
+      );
+    }
+    return results.reduce(
+      (acc, r, i) =>
+        Object.assign(
+          acc,
+          matchesToLoad[i].shouldCallHandler()
+            ? { [matchesToLoad[i].route.id]: r }
+            : {},
+        ),
+      {},
+    );
+  }
+
   describe("loaders", () => {
     it("should allow a custom implementation to passthrough to default behavior", async () => {
       let dataStrategy = mockDataStrategy(({ matches }) =>
@@ -568,6 +589,112 @@ describe("router dataStrategy", () => {
           route: expect.objectContaining({ id: "child" }),
         }),
       ]);
+      expect(t.router.state.actionData).toMatchObject({
+        child: "CHILD ACTION",
+      });
+      expect(t.router.state.loaderData).toMatchObject({
+        root: "ROOT",
+        parent: "PARENT",
+        child: "CHILD",
+      });
+    });
+
+    it("indicates which routes need to load via match.shouldCallHandler()", async () => {
+      let shouldCallHandlerValues: Record<string, boolean>[] = [];
+      let dataStrategy = jest.fn<
+        ReturnType<DataStrategyFunction>,
+        Parameters<DataStrategyFunction>
+      >(async ({ request, matches }) => {
+        let values = matches.reduce(
+          (acc, m) =>
+            Object.assign(acc, {
+              [m.route.id]: m.shouldCallHandler(),
+            }),
+          {},
+        );
+        shouldCallHandlerValues.push(values);
+        let matchesToLoad = matches.filter((m) => m.shouldCallHandler());
+        let results = await Promise.all(matchesToLoad.map((m) => m.resolve()));
+        return keyedResultsUsingShouldCallHandler(matchesToLoad, results);
+      });
+      let [lazy, lazyDeferred] = createAsyncStub();
+      let t = setup({
+        routes: [
+          {
+            id: "root",
+            path: "/",
+            loader: true,
+            children: [
+              {
+                id: "parent",
+                path: "parent",
+                loader: true,
+                action: true,
+                children: [
+                  {
+                    id: "child",
+                    path: "child",
+                    lazy,
+                  },
+                ],
+              },
+            ],
+          },
+        ],
+        dataStrategy,
+        hydrationData: {
+          // don't call dataStrategy on hydration
+          loaderData: { root: null },
+        },
+      });
+
+      let A = await t.navigate("/");
+      expect(shouldCallHandlerValues[0]).toEqual({ root: true });
+      await A.loaders.root.resolve("ROOT");
+      expect(t.router.state.loaderData).toMatchObject({
+        root: "ROOT",
+      });
+
+      let B = await t.navigate("/parent");
+      expect(shouldCallHandlerValues[1]).toEqual({ root: false, parent: true });
+      await B.loaders.parent.resolve("PARENT");
+      expect(t.router.state.loaderData).toMatchObject({
+        root: "ROOT",
+        parent: "PARENT",
+      });
+
+      await t.navigate("/parent/child");
+      expect(shouldCallHandlerValues[2]).toEqual({
+        root: false,
+        parent: false,
+        child: true,
+      });
+      await lazyDeferred.resolve({
+        action: () => "CHILD ACTION",
+        loader: () => "CHILD",
+        shouldRevalidate: () => false,
+      });
+      expect(t.router.state.loaderData).toMatchObject({
+        root: "ROOT",
+        parent: "PARENT",
+        child: "CHILD",
+      });
+
+      await t.navigate("/parent/child", {
+        formMethod: "post",
+        formData: createFormData({}),
+      });
+      await tick();
+      expect(shouldCallHandlerValues[3]).toEqual({
+        root: false,
+        parent: false,
+        child: true,
+      });
+      expect(shouldCallHandlerValues[4]).toEqual({
+        root: true,
+        parent: true,
+        child: false,
+      });
       expect(t.router.state.actionData).toMatchObject({
         child: "CHILD ACTION",
       });
@@ -1295,6 +1422,109 @@ describe("router dataStrategy", () => {
       });
     });
 
+    it("allows middleware/context implementations when some routes don't need to revalidate (using shouldCallHandler)", async () => {
+      let t = setup({
+        routes: [
+          {
+            path: "/",
+          },
+          {
+            id: "parent",
+            path: "/parent",
+            loader: true,
+            handle: {
+              context: {
+                parent: () => ({ id: "parent" }),
+              },
+              middleware(context) {
+                context.parent.whatever = "PARENT MIDDLEWARE";
+              },
+            },
+            children: [
+              {
+                id: "child",
+                path: "child",
+                loader: true,
+                handle: {
+                  context: {
+                    child: () => ({ id: "child" }),
+                  },
+                  middleware(context) {
+                    context.child.whatever = "CHILD MIDDLEWARE";
+                  },
+                },
+              },
+            ],
+          },
+        ],
+        async dataStrategy({ matches }) {
+          // Run context/middleware sequentially
+          let context = matches.reduce((acc, m) => {
+            if (m.route.handle?.context) {
+              let matchContext = Object.entries(m.route.handle.context).reduce(
+                (acc, [key, value]) =>
+                  Object.assign(acc, {
+                    // @ts-expect-error
+                    [key]: value(),
+                  }),
+                {},
+              );
+              Object.assign(acc, matchContext);
+            }
+            if (m.route.handle?.middleware) {
+              m.route.handle.middleware(acc);
+            }
+            return acc;
+          }, {});
+
+          // Run loaders in parallel only exposing contexts from above
+          let matchesToLoad = matches.filter((m) => m.shouldCallHandler());
+          let results = await Promise.all(
+            matchesToLoad.map((m) =>
+              m.resolve((callHandler) => callHandler(context)),
+            ),
+          );
+          return keyedResultsUsingShouldCallHandler(matchesToLoad, results);
+        },
+      });
+
+      let A = await t.navigate("/parent");
+      await A.loaders.parent.resolve("PARENT");
+      expect(t.router.state.navigation.state).toBe("idle");
+      expect(t.router.state.loaderData).toMatchObject({
+        parent: "PARENT",
+      });
+
+      let B = await t.navigate("/parent/child");
+
+      // Loaders are called with context from their level and above, and
+      // context reflects any values set by middleware
+      expect(B.loaders.child.stub).toHaveBeenCalledWith(
+        expect.objectContaining({
+          request: expect.any(Request),
+          params: expect.any(Object),
+        }),
+        {
+          parent: {
+            id: "parent",
+            whatever: "PARENT MIDDLEWARE",
+          },
+          child: {
+            id: "child",
+            whatever: "CHILD MIDDLEWARE",
+          },
+        },
+      );
+
+      await B.loaders.child.resolve("CHILD");
+      expect(t.router.state.navigation.state).toBe("idle");
+
+      expect(t.router.state.loaderData).toMatchObject({
+        parent: "PARENT",
+        child: "CHILD",
+      });
+    });
+
     it("allows automatic caching of loader results", async () => {
       let cache: Record<string, unknown> = {};
       let t = setup({
@@ -1404,6 +1634,187 @@ describe("router dataStrategy", () => {
         loaderData: {
           parent: "PARENT**",
           child: "CHILD**",
+        },
+      });
+    });
+
+    it("allows automatic caching of loader results (using shouldCallHandler)", async () => {
+      let cache: Record<string, unknown> = {};
+      let t = setup({
+        routes: [
+          {
+            path: "/",
+          },
+          {
+            id: "parent",
+            path: "/parent",
+            loader: true,
+            handle: {
+              cacheKey: (url: string) => new URL(url).pathname,
+            },
+            children: [
+              {
+                id: "child",
+                path: "child",
+                loader: true,
+                action: true,
+              },
+            ],
+          },
+        ],
+        async dataStrategy({ request, matches }) {
+          const getCacheKey = (m: DataStrategyMatch) =>
+            m.route.handle?.cacheKey
+              ? [m.route.id, m.route.handle.cacheKey(request.url)].join("-")
+              : null;
+
+          if (request.method !== "GET") {
+            // invalidate on actions
+            cache = {};
+          }
+
+          let matchesToLoad = matches.filter((m) => m.shouldCallHandler());
+          let results = await Promise.all(
+            matchesToLoad.map(async (m) => {
+              return m.resolve(async (handler) => {
+                let key = getCacheKey(m);
+                if (key && cache[key]) {
+                  return cache[key];
+                }
+
+                let dsResult = await handler();
+                if (key && request.method === "GET") {
+                  cache[key] = dsResult;
+                }
+
+                return dsResult;
+              });
+            }),
+          );
+          return keyedResultsUsingShouldCallHandler(matchesToLoad, results);
+        },
+      });
+
+      let A = await t.navigate("/parent/child");
+      await A.loaders.parent.resolve("PARENT");
+      await A.loaders.child.resolve("CHILD");
+
+      expect(t.router.state).toMatchObject({
+        navigation: { state: "idle" },
+        loaderData: {
+          parent: "PARENT",
+          child: "CHILD",
+        },
+      });
+
+      // Changing search params should force revalidation, but pathname-based
+      // cache will serve the old data
+      let B = await t.navigate("/parent/child?a=b");
+      await B.loaders.child.resolve("CHILD*");
+
+      expect(t.router.state).toMatchObject({
+        navigation: { state: "idle" },
+        loaderData: {
+          parent: "PARENT",
+          child: "CHILD*",
+        },
+      });
+
+      // Useless resolution - handler was never called for parent
+      await B.loaders.parent.resolve("PARENT*");
+
+      expect(t.router.state).toMatchObject({
+        navigation: { state: "idle" },
+        loaderData: {
+          parent: "PARENT",
+          child: "CHILD*",
+        },
+      });
+
+      // Action to invalidate the cache
+      let C = await t.navigate("/parent/child?a=b", {
+        formMethod: "post",
+        formData: createFormData({}),
+      });
+      await C.actions.child.resolve("ACTION");
+      await C.loaders.parent.resolve("PARENT**");
+      await C.loaders.child.resolve("CHILD**");
+
+      expect(t.router.state).toMatchObject({
+        navigation: { state: "idle" },
+        actionData: {
+          child: "ACTION",
+        },
+        loaderData: {
+          parent: "PARENT**",
+          child: "CHILD**",
+        },
+      });
+    });
+
+    it("allows defining a custom revalidation behavior", async () => {
+      let t = setup({
+        routes: [
+          {
+            id: "index",
+            path: "/",
+            action: true,
+            loader: true,
+          },
+        ],
+        async dataStrategy({ request, matches, context }) {
+          if (request.method !== "GET") {
+            context.actionMethod = request.method;
+          }
+          // Don't revalidate on PUT requests
+          let defaultShouldRevalidate =
+            context.actionMethod === "PUT" ? false : undefined;
+          let matchesToLoad = matches.filter((m) =>
+            m.shouldCallHandler(defaultShouldRevalidate),
+          );
+          let results = await Promise.all(
+            matchesToLoad.map((m) => m.resolve()),
+          );
+          return keyedResultsUsingShouldCallHandler(matchesToLoad, results);
+        },
+        hydrationData: {
+          loaderData: {
+            index: "INDEX",
+          },
+        },
+      });
+
+      let A = await t.navigate("/", {
+        formMethod: "post",
+        formData: createFormData({}),
+      });
+      await A.actions.index.resolve("ACTION1");
+      await A.loaders.index.resolve("INDEX1");
+      expect(t.router.state.navigation.state).toBe("idle");
+      expect(t.router.state).toMatchObject({
+        navigation: { state: "idle" },
+        actionData: {
+          index: "ACTION1",
+        },
+        loaderData: {
+          index: "INDEX1",
+        },
+      });
+
+      let B = await t.navigate("/", {
+        formMethod: "put",
+        formData: createFormData({}),
+      });
+      await B.actions.index.resolve("ACTION2");
+      //await B.loaders.index.resolve("INDEX2"); // no-op
+      expect(t.router.state.navigation.state).toBe("idle");
+      expect(t.router.state).toMatchObject({
+        navigation: { state: "idle" },
+        actionData: {
+          index: "ACTION2",
+        },
+        loaderData: {
+          index: "INDEX1",
         },
       });
     });