+++ packages/react-router/lib/context.ts
@@ -87,8 +87,9 @@ export type PatchRoutesOnNavigationFunction =
 
 export interface DataRouterContextObject
   // Omit `future` since those can be pulled from the `router`
-  // `NavigationContext` needs future since it doesn't have a `router` in all cases
-  extends Omit<NavigationContextObject, "future"> {
+  // `NavigationContext` needs `future`/`unstable_useTransitions` since it doesn't
+  // have a `router` in all cases
+  extends Omit<NavigationContextObject, "future" | "unstable_useTransitions"> {
   router: Router;
   staticContext?: StaticHandlerContext;
   unstable_onError?: unstable_ClientOnErrorFunction;
@@ -180,6 +181,7 @@ interface NavigationContextObject {
   basename: string;
   navigator: Navigator;
   static: boolean;
+  unstable_useTransitions: boolean;
   // TODO: Re-introduce a singular `FutureConfig` once we land our first
   // future.unstable_ or future.v8_ flag
   future: {};